Implemented "Step 10" action = signup
Added some checks to the server side for user input. Specifically there
are checks to determine if the user input is blank/null, and that the passwords
match. The rest of the checks (from client side) should probably be implemented
as well. 

Added Error displays to signup.jsp when user info is invalid when it reaches
the server.