Fixed a few bugs that caused the application to not allow new users to be added.
membershipServlet.java:
`Added salt and hashed the "confirm" password string. The password was
``being hashed, but not the confirm field. This meant when they reached
``the server they did not match and were returning an error.

UserDB.java:
`Added a ninth field (?) to the prepared statement for insertUser
``because we are now adding salt to the insert. Before adding this new
``sign-ups would fail, due to an SQL Exception